Basic styling to have a non-plain ui
